update encoder and make universal
